+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Version 1.18
 ============
 - Add function to switch current viewport to the relative camera mode with options to control the position of the camera
 - Add function to display debug lines in the environment
+- Add function to send Lua commands to be executed inside the simulation
 
 Version 1.17
 ============